+++ src/split_pattern_driver.h
@@ -12,7 +12,7 @@ enum DataType {RELWGHT, ABSWGHT, SCOPE};
 
 typedef struct Elmt{
   SP::DataType      type;
-  float             value; // Weight or # of scope if it is repeated
+  double             value; // Weight or # of scope if it is repeated
   std::string       actions;
   std::list<Elmt*>  subElmts;
 } Elmt;
@@ -29,10 +29,10 @@ public:
   void exitSubScope() {currentScope.pop();}
   void wasConstScope();
 
-  void computePattern(float _totalLength); /* Find the optimal repetitions,
+  void computePattern(double _totalLength); /* Find the optimal repetitions,
                                             * then compute final vectors. */
 
-  const std::vector<float>&        getWeights() const {return finalWeights;}
+  const std::vector<double>&        getWeights() const {return finalWeights;}
   const std::vector<std::string>&  getActions() const {return finalActions;}
 
 private:
@@ -47,7 +47,7 @@ private:
   SP::SP_Scanner *scanner = nullptr;
 
   bool hasRelWeight;
-  float totalLength;
+  double totalLength;
 
   std::stack<std::list<Elmt*>*> currentScope;
   std::list<Elmt*> pattern;
@@ -55,10 +55,10 @@ private:
                                 // -1 if the scope does not have a star -> not to be repeated
 
   std::list<Elmt*> patternInstance;
-  float totalRelWeight;
-  float totalAbsLength;
+  double totalRelWeight;
+  double totalAbsLength;
 
-  std::vector<float> finalWeights;
+  std::vector<double> finalWeights;
   std::vector<std::string> finalActions;
 };
